New style for Chicago 17th (author-date) that uses "no." for journal issues
Chicago Manual of Style 17th edition (author-date) - Turabian journal citation
So instead of: "Random Journal 5 (1): 123-321."
You will see: "Random Journal 5, no. 1: 123-321."
It is the standard for Turabian 9th, which is for my PhD program. I don't have the time or knowledge to write up a citation for Turabian 9th edition author-date style. So for now, this will have to suffice.